Disable obsolete HTCondor systemd startup fix
A workaround for a bug in cloud-init wrt hostname initialisation
that affected HTCondor has been disabled; the underlying issue is
resolved in recent CentOS 7 versions and the fix now does more
bad than good in that it prevents a clean termination of the
HTCondor-service during system shutdown.

Build version bumped to 24.