Change autodrain threshold to 47h
Having VG/NEMO CN nodes autodrain after only 24h has unwanted
consequences in that ROCED by default counts unclaimed slots of
draining hosts as available and thus often fails to start new
instances when they are required. (There is a config option named
'ignore_draining_machines' which does the opposite of what you'd
expect it to: when set to /false/ ROCED does /not/ count slots
of draining hosts as available; but with this option activted,
it also fails to count claimed slots on draining instances as
running and thus will often start unneeded instances, which is
not exactly desireable either...)

Thus this commit just sets the autodrain threshold to 47h and
the grace period to 30m (VMs on NEMO can run up to 48h walltime)
while keeping the default value of 'ignore_draining_machines'
unchanged which could cause delayed scheduling of jobs in some
cases but will hopefully alleviates the problem to the point
where it is no longer really one.

(Sry for the lengthy commit message, but I lack the time to make
it shorter...)

Bumped image version to 25.